Apply a function with progress bars.

Apply a function with progress bars.

with_progress(fun, total, ...)

Arguments

fun
The function to be apply

total
The total number of elements to be mapped. If omitted an attempt will be made to infer the correct number.

...
Arguments passed on to progress_bar

total

the total number of elements

title

the title of the progress bar

type

the type of progress bar to create as a string, or an R6ClassGenerator object for a class that inherits from the "R6 Progress Base Class".

Examples

# with purrr functions
long_function <- function(x, how.long=0.05){
    Sys.sleep(how.long)
    x
}
# \donttest{
purrr::walk(1:100, with_progress(long_function))
purrr::walk2(1:100, 0.01, with_progress(long_function))
# }

# with dplyr::group_map
# \donttest{
if(require(dplyr)){
group_function <- function(x, y, how.long=0.05){
    Sys.sleep(how.long)
    x
}
group_map( group_by(mtcars, cyl, gear)
         , with_progress(group_function, type='line')
         , how.long=1/3)
group_walk( group_by_all(mtcars)
          , with_progress(group_function, type='box')
          , how.long=1)
}

#> Loading required package: dplyr

#> Warning: there is no package called ‘dplyr’

# }
# with standard apply functions
sapply(1:100, with_progress(long_function, type='txt'), 0.001)

#>   0/100|                                              |(0%) NA remaining  0/100|                                              |(0%) NA remaining  1/100|                                              |(1%) 00:00:00 remaining  1/100|                                              |(1%) 00:00:01 remaining  2/100|                                              |(2%) 00:00:00 remaining  2/100|                                              |(2%) 00:00:00 remaining  3/100|=                                             |(3%) 00:00:00 remaining  3/100|=                                             |(3%) 00:00:01 remaining  4/100|=                                             |(4%) 00:00:00 remaining  4/100|=                                             |(4%) 00:00:00 remaining  5/100|==                                            |(5%) 00:00:00 remaining  5/100|==                                            |(5%) 00:00:00 remaining  6/100|==                                            |(6%) 00:00:00 remaining  6/100|==                                            |(6%) 00:00:00 remaining  7/100|===                                           |(7%) 00:00:00 remaining  7/100|===                                           |(7%) 00:00:01 remaining  8/100|===                                           |(8%) 00:00:01 remaining  8/100|===                                           |(8%) 00:00:01 remaining  9/100|====                                          |(9%) 00:00:00 remaining  9/100|====                                          |(9%) 00:00:01 remaining 10/100|====                                          |(10%) 00:00:00 remaining 10/100|====                                          |(10%) 00:00:01 remaining 11/100|=====                                         |(11%) 00:00:00 remaining 11/100|=====                                         |(11%) 00:00:00 remaining 12/100|=====                                         |(12%) 00:00:00 remaining 12/100|=====                                         |(12%) 00:00:01 remaining 13/100|=====                                         |(13%) 00:00:00 remaining 13/100|=====                                         |(13%) 00:00:00 remaining 14/100|======                                        |(14%) 00:00:00 remaining 14/100|======                                        |(14%) 00:00:00 remaining 15/100|======                                        |(15%) 00:00:00 remaining 15/100|======                                        |(15%) 00:00:00 remaining 16/100|=======                                       |(16%) 00:00:00 remaining 16/100|=======                                       |(16%) 00:00:00 remaining 17/100|=======                                       |(17%) 00:00:00 remaining 17/100|=======                                       |(17%) 00:00:00 remaining 18/100|========                                      |(18%) 00:00:00 remaining 18/100|========                                      |(18%) 00:00:00 remaining 19/100|========                                      |(19%) 00:00:00 remaining 19/100|========                                      |(19%) 00:00:00 remaining 20/100|=========                                     |(20%) 00:00:00 remaining 20/100|=========                                     |(20%) 00:00:00 remaining 21/100|=========                                     |(21%) 00:00:00 remaining 21/100|=========                                     |(21%) 00:00:00 remaining 22/100|==========                                    |(22%) 00:00:00 remaining 22/100|==========                                    |(22%) 00:00:00 remaining 23/100|==========                                    |(23%) 00:00:00 remaining 23/100|==========                                    |(23%) 00:00:00 remaining 24/100|===========                                   |(24%) 00:00:00 remaining 24/100|===========                                   |(24%) 00:00:00 remaining 25/100|===========                                   |(25%) 00:00:00 remaining 25/100|===========                                   |(25%) 00:00:00 remaining 26/100|===========                                   |(26%) 00:00:00 remaining 26/100|===========                                   |(26%) 00:00:00 remaining 27/100|============                                  |(27%) 00:00:00 remaining 27/100|============                                  |(27%) 00:00:00 remaining 28/100|============                                  |(28%) 00:00:00 remaining 28/100|============                                  |(28%) 00:00:00 remaining 29/100|=============                                 |(28%) 00:00:00 remaining 29/100|=============                                 |(28%) 00:00:00 remaining 30/100|=============                                 |(30%) 00:00:00 remaining 30/100|=============                                 |(30%) 00:00:00 remaining 31/100|==============                                |(31%) 00:00:00 remaining 31/100|==============                                |(31%) 00:00:00 remaining 32/100|==============                                |(32%) 00:00:00 remaining 32/100|==============                                |(32%) 00:00:00 remaining 33/100|===============                               |(33%) 00:00:00 remaining 33/100|===============                               |(33%) 00:00:00 remaining 34/100|===============                               |(34%) 00:00:00 remaining 34/100|===============                               |(34%) 00:00:00 remaining 35/100|================                              |(35%) 00:00:00 remaining 35/100|================                              |(35%) 00:00:00 remaining 36/100|================                              |(36%) 00:00:00 remaining 36/100|================                              |(36%) 00:00:00 remaining 37/100|=================                             |(37%) 00:00:00 remaining 37/100|=================                             |(37%) 00:00:00 remaining 38/100|=================                             |(38%) 00:00:00 remaining 38/100|=================                             |(38%) 00:00:00 remaining 39/100|=================                             |(39%) 00:00:00 remaining 39/100|=================                             |(39%) 00:00:00 remaining 40/100|==================                            |(40%) 00:00:00 remaining 40/100|==================                            |(40%) 00:00:00 remaining 41/100|==================                            |(41%) 00:00:00 remaining 41/100|==================                            |(41%) 00:00:00 remaining 42/100|===================                           |(42%) 00:00:00 remaining 42/100|===================                           |(42%) 00:00:00 remaining 43/100|===================                           |(43%) 00:00:00 remaining 43/100|===================                           |(43%) 00:00:00 remaining 44/100|====================                          |(44%) 00:00:00 remaining 44/100|====================                          |(44%) 00:00:00 remaining 45/100|====================                          |(45%) 00:00:00 remaining 45/100|====================                          |(45%) 00:00:00 remaining 46/100|=====================                         |(46%) 00:00:00 remaining 46/100|=====================                         |(46%) 00:00:00 remaining 47/100|=====================                         |(47%) 00:00:00 remaining 47/100|=====================                         |(47%) 00:00:00 remaining 48/100|======================                        |(48%) 00:00:00 remaining 48/100|======================                        |(48%) 00:00:00 remaining 49/100|======================                        |(49%) 00:00:00 remaining 49/100|======================                        |(49%) 00:00:00 remaining 50/100|=======================                       |(50%) 00:00:00 remaining 50/100|=======================                       |(50%) 00:00:00 remaining 51/100|=======================                       |(51%) 00:00:00 remaining 51/100|=======================                       |(51%) 00:00:00 remaining 52/100|=======================                       |(52%) 00:00:00 remaining 52/100|=======================                       |(52%) 00:00:00 remaining 53/100|========================                      |(53%) 00:00:00 remaining 53/100|========================                      |(53%) 00:00:00 remaining 54/100|========================                      |(54%) 00:00:00 remaining 54/100|========================                      |(54%) 00:00:00 remaining 55/100|=========================                     |(55%) 00:00:00 remaining 55/100|=========================                     |(55%) 00:00:00 remaining 56/100|=========================                     |(56%) 00:00:00 remaining 56/100|=========================                     |(56%) 00:00:00 remaining 57/100|==========================                    |(56%) 00:00:00 remaining 57/100|==========================                    |(56%) 00:00:00 remaining 58/100|==========================                    |(57%) 00:00:00 remaining 58/100|==========================                    |(57%) 00:00:00 remaining 59/100|===========================                   |(59%) 00:00:00 remaining 59/100|===========================                   |(59%) 00:00:00 remaining 60/100|===========================                   |(60%) 00:00:00 remaining 60/100|===========================                   |(60%) 00:00:00 remaining 61/100|============================                  |(61%) 00:00:00 remaining 61/100|============================                  |(61%) 00:00:00 remaining 62/100|============================                  |(62%) 00:00:00 remaining 62/100|============================                  |(62%) 00:00:00 remaining 63/100|============================                  |(63%) 00:00:00 remaining 63/100|============================                  |(63%) 00:00:00 remaining 64/100|=============================                 |(64%) 00:00:00 remaining 64/100|=============================                 |(64%) 00:00:00 remaining 65/100|=============================                 |(65%) 00:00:00 remaining 65/100|=============================                 |(65%) 00:00:00 remaining 66/100|==============================                |(66%) 00:00:00 remaining 66/100|==============================                |(66%) 00:00:00 remaining 67/100|==============================                |(67%) 00:00:00 remaining 67/100|==============================                |(67%) 00:00:00 remaining 68/100|===============================               |(68%) 00:00:00 remaining 68/100|===============================               |(68%) 00:00:00 remaining 69/100|===============================               |(69%) 00:00:00 remaining 69/100|===============================               |(69%) 00:00:00 remaining 70/100|================================              |(70%) 00:00:00 remaining 70/100|================================              |(70%) 00:00:00 remaining 71/100|================================              |(71%) 00:00:00 remaining 71/100|================================              |(71%) 00:00:00 remaining 72/100|=================================             |(72%) 00:00:00 remaining 72/100|=================================             |(72%) 00:00:00 remaining 73/100|=================================             |(73%) 00:00:00 remaining 73/100|=================================             |(73%) 00:00:00 remaining 74/100|==================================            |(74%) 00:00:00 remaining 74/100|==================================            |(74%) 00:00:00 remaining 75/100|==================================            |(75%) 00:00:00 remaining 75/100|==================================            |(75%) 00:00:00 remaining 76/100|==================================            |(76%) 00:00:00 remaining 76/100|==================================            |(76%) 00:00:00 remaining 77/100|===================================           |(77%) 00:00:00 remaining 77/100|===================================           |(77%) 00:00:00 remaining 78/100|===================================           |(78%) 00:00:00 remaining 78/100|===================================           |(78%) 00:00:00 remaining 79/100|====================================          |(79%) 00:00:00 remaining 79/100|====================================          |(79%) 00:00:00 remaining 80/100|====================================          |(80%) 00:00:00 remaining 80/100|====================================          |(80%) 00:00:00 remaining 81/100|=====================================         |(81%) 00:00:00 remaining 81/100|=====================================         |(81%) 00:00:00 remaining 82/100|=====================================         |(82%) 00:00:00 remaining 82/100|=====================================         |(82%) 00:00:00 remaining 83/100|======================================        |(83%) 00:00:00 remaining 83/100|======================================        |(83%) 00:00:00 remaining 84/100|======================================        |(84%) 00:00:00 remaining 84/100|======================================        |(84%) 00:00:00 remaining 85/100|=======================================       |(85%) 00:00:00 remaining 85/100|=======================================       |(85%) 00:00:00 remaining 86/100|=======================================       |(86%) 00:00:00 remaining 86/100|=======================================       |(86%) 00:00:00 remaining 87/100|========================================      |(87%) 00:00:00 remaining 87/100|========================================      |(87%) 00:00:00 remaining 88/100|========================================      |(88%) 00:00:00 remaining 88/100|========================================      |(88%) 00:00:00 remaining 89/100|========================================      |(89%) 00:00:00 remaining 89/100|========================================      |(89%) 00:00:00 remaining 90/100|=========================================     |(90%) 00:00:00 remaining 90/100|=========================================     |(90%) 00:00:00 remaining 91/100|=========================================     |(91%) 00:00:00 remaining 91/100|=========================================     |(91%) 00:00:00 remaining 92/100|==========================================    |(92%) 00:00:00 remaining 92/100|==========================================    |(92%) 00:00:00 remaining 93/100|==========================================    |(93%) 00:00:00 remaining 93/100|==========================================    |(93%) 00:00:00 remaining 94/100|===========================================   |(94%) 00:00:00 remaining 94/100|===========================================   |(94%) 00:00:00 remaining 95/100|===========================================   |(95%) 00:00:00 remaining 95/100|===========================================   |(95%) 00:00:00 remaining 96/100|============================================  |(96%) 00:00:00 remaining 96/100|============================================  |(96%) 00:00:00 remaining 97/100|============================================  |(97%) 00:00:00 remaining 97/100|============================================  |(97%) 00:00:00 remaining 98/100|============================================= |(98%) 00:00:00 remaining 98/100|============================================= |(98%) 00:00:00 remaining 99/100|============================================= |(99%) 00:00:00 remaining 99/100|============================================= |(99%) 00:00:00 remaining 100/100|==============================================|(100%) 00:00:00 remainin

#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#>  [19]  19  20  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36
#>  [37]  37  38  39  40  41  42  43  44  45  46  47  48  49  50  51  52  53  54
#>  [55]  55  56  57  58  59  60  61  62  63  64  65  66  67  68  69  70  71  72
#>  [73]  73  74  75  76  77  78  79  80  81  82  83  84  85  86  87  88  89  90
#>  [91]  91  92  93  94  95  96  97  98  99 100